Move common models into common location for import
